+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to
 - ✨(frontend) add word count to doc header toolbox #2549
 - ✨(frontend) add find and replace feature to the editor #2570
 - ✨(frontend) add math and diagram blocks to the editor #2617
+- ✨(frontend) warn the user on too big file uploads #2522
 
 ### Changed